+++ CHANGELOG.md
@@ -17,7 +17,8 @@ this release does not alter the usage of the API under normal circumstances.
 
 ### Added
 
-- Added `ContainerAware` interface, which is implemented by `Container` itself
+- Added `Catalog` which is a collection of dependency declarations that can be reused by multiple container instances
+- Added `ContainerAware` interface, which is implemented `Container` and `Catalog`
 
 ### Changed